+++ dymension/libs/kaspa/lib/relayer/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod deposit;
-pub mod hub_to_kaspa_builder;
 pub mod withdraw;
 pub mod confirmation;
+pub mod hub_to_kaspa;
 
 // Re-export the main function for easier access
-pub use hub_to_kaspa_builder::build_kaspa_withdrawal_pskts;
+pub use hub_to_kaspa::build_withdrawal_pskts;
